Refractor of maps.ts functions
You should now create maps via the class in TraversableMap.ts instead of via the function in maps.ts. This was made in an effort to reduce code, increase readability and modularity. Though not 100% tested, it works fine for the the small tests I did and is probably enough for what we're accomplishing.